+++ src/main/java/hello/cluebackend/domain/classroomuser/service/ClassroomUserService.java
@@ -1,7 +1,7 @@
 package hello.cluebackend.domain.classroomuser.service;
 
 import hello.cluebackend.domain.classroom.model.ClassRoom;
-import hello.cluebackend.domain.classroom.service.ClassRoomCommandService;
+import hello.cluebackend.domain.classroom.service.ClassRoomQueryService;
 import hello.cluebackend.domain.classroomuser.model.ClassRoomUser;
 import hello.cluebackend.infrastructure.persistence.classroomuser.ClassRoomUserJpaRepository;
 import hello.cluebackend.domain.user.model.UserEntity;
@@ -17,7 +17,7 @@ import java.util.UUID;
 public class ClassroomUserService {
   private final ClassRoomUserJpaRepository classRoomUserJpaRepository;
   private final UserService userService;
-  private final ClassRoomCommandService classRoomCommandService;
+  private final ClassRoomQueryService classRoomQueryService;
 
   // 해당 교실에 속한 모든 사용자 조회
   public List<UserEntity> findAllClassroomUser(ClassRoom classRoom){
@@ -30,7 +30,7 @@ public class ClassroomUserService {
 
   // 수업실에 해당 유저가 속하는지 확인하는 로직
   public boolean isUserInClassroom(UUID classRoomId, UUID userId) {
-    ClassRoom classRoom = classRoomCommandService.findById(userId, classRoomId).toEntity();
+    ClassRoom classRoom = classRoomQueryService.findById(userId, classRoomId).toEntity();
     UserEntity user = userService.findById(userId).toEntity();
 
     return classRoomUserJpaRepository.existsByClassRoomAndUser(classRoom,user);
