+++ src/main/java/hello/cluebackend/domain/submission/service/SubmissionQueryService.java
@@ -8,7 +8,7 @@ import hello.cluebackend.domain.submission.model.Submission;
 import hello.cluebackend.domain.submission.model.SubmissionAttachment;
 import hello.cluebackend.infrastructure.persistence.submission.SubmissionJpaRepository;
 import hello.cluebackend.domain.classroom.model.ClassRoom;
-import hello.cluebackend.domain.classroom.service.ClassRoomCommandService;
+import hello.cluebackend.domain.classroom.service.ClassRoomQueryService;
 import hello.cluebackend.domain.classroomuser.service.ClassroomUserService;
 import hello.cluebackend.infrastructure.persistence.submissionattachment.SubmissionAttachmentJpaRepository;
 import hello.cluebackend.domain.user.model.UserEntity;
@@ -26,14 +26,14 @@ public class SubmissionQueryService {
   private final SubmissionJpaRepository submissionJpaRepository;
   private final SubmissionAttachmentJpaRepository submissionAttachmentJpaRepository;
   private final ClassroomUserService classroomUserService;
-  private final ClassRoomCommandService classRoomCommandService;
+  private final ClassRoomQueryService classRoomQueryService;
   private final FileService fileService;
   private final SubmissionCommandService submissionCommandService;
 
   // 해당 교실 모든 학생에게 과제 부여 & 제출 과제 생성
   @Transactional
   public void assignToAllStudentsInClassroom(UUID userId, UUID classroomId, Assignment assignment){
-    ClassRoom classRoom = classRoomCommandService.findById(userId ,classroomId).toEntity();
+    ClassRoom classRoom = classRoomQueryService.findById(userId ,classroomId).toEntity();
     List<UserEntity> users = classroomUserService.findAllClassroomUser(classRoom);
     List<Submission> submissions = users.stream()
             .map(u -> new Submission(assignment, u,assignment.getClassRoom(), false, null))
